Atualizações Summary

+++ MarketDataFinance.Application/ViewModels/QuoteViewModel.cs
@@ -12,7 +12,7 @@ namespace MarketDataFinance.Application.ViewModels
         public string Exchange { get; set; }
         public double FiftyTwoWeekHighChange { get; set; }
         public string CustomPriceAlertConfidence { get; set; }
-        public int AverageDailyVolume10Day { get; set; }
+        public long AverageDailyVolume10Day { get; set; }
         public string ShortName { get; set; }
         public bool Triggerable { get; set; }
         public double AnnualReturnNavY3 { get; set; }
@@ -46,7 +46,7 @@ namespace MarketDataFinance.Application.ViewModels
         public string FullExchangeName { get; set; }
         public double RegularMarketPreviousClose { get; set; }
         public string Symbol { get; set; }
-        public int RegularMarketVolume { get; set; }
+        public long RegularMarketVolume { get; set; }
         public double YtdReturn { get; set; }
         public string Market { get; set; }
         public double FiftyDayAverage { get; set; }
@@ -59,7 +59,7 @@ namespace MarketDataFinance.Application.ViewModels
         public int PriceHint { get; set; }
         public int AskSize { get; set; }
         public string TypeDisp { get; set; }
-        public int AverageDailyVolume3Month { get; set; }
+        public long AverageDailyVolume3Month { get; set; }
         public string QuoteType { get; set; }
         public string RegularMarketDayRange { get; set; }
         public double FiftyTwoWeekHighChangePercent { get; set; }
@@ -74,8 +74,8 @@ namespace MarketDataFinance.Application.ViewModels
         public double FiftyTwoWeekHigh { get; set; }
         public long NetAssets { get; set; }
         public bool EsgPopulated { get; set; }
-        public int TrailingAnnualDividendYield { get; set; }
-        public int TrailingAnnualDividendRate { get; set; }
+        public double TrailingAnnualDividendYield { get; set; }
+        public double TrailingAnnualDividendRate { get; set; }
         public int DividendDate { get; set; }
         public double EpsTrailingTwelveMonths { get; set; }
         public double TrailingPE { get; set; }
